+++ Src/AwesomeAssertions/Primitives/StringEqualityStrategy.cs
@@ -29,6 +29,11 @@ internal class StringEqualityStrategy : IStringComparisonStrategy
 
     public void ValidateAgainstMismatch(AssertionChain assertionChain, string subject, string expected)
     {
+        if (comparer.Equals(subject, expected))
+        {
+            return;
+        }
+
         if (ValidateAgainstSuperfluousWhitespace(assertionChain, subject, expected))
         {
             return;
@@ -189,7 +194,7 @@ internal class StringEqualityStrategy : IStringComparisonStrategy
     }
 
     /// <summary>
-    /// Get index of the first mismatch between <paramref name="subject"/> and <paramref name="expected"/>. 
+    /// Get index of the first mismatch between <paramref name="subject"/> and <paramref name="expected"/>.
     /// </summary>
     /// <param name="subject"></param>
     /// <param name="expected"></param>